Change matlab MEX files to use unsigned/size_t as appropriate

+++ matlab/mex/socket.c
@@ -15,7 +15,8 @@
 #include "socket.h"
 
 
-mxArray *create_double_array(int rows, int cols, double **reals, double **imags)
+mxArray *create_double_array(
+    size_t rows, size_t cols, double **reals, double **imags)
 {
     mxComplexity type = imags ? mxCOMPLEX : mxREAL;
     mwSize dims[2] = { rows, cols };
@@ -26,7 +27,8 @@ mxArray *create_double_array(int rows, int cols, double **reals, double **imags)
     return array;
 }
 
-mxArray *create_single_array(int rows, int cols, float **reals, float **imags)
+mxArray *create_single_array(
+    size_t rows, size_t cols, float **reals, float **imags)
 {
     mxComplexity type = imags ? mxCOMPLEX : mxREAL;
     mwSize dims[2] = { rows, cols };
@@ -52,7 +54,7 @@ int connect_server(const char *hostname, int port)
         .sin_family = AF_INET,
         .sin_port = htons(port),
     };
-    memcpy(&s_in.sin_addr.s_addr, hostent->h_addr, hostent->h_length);
+    memcpy(&s_in.sin_addr.s_addr, hostent->h_addr, (size_t) hostent->h_length);
 
     TEST_SOCK_(sock,
         connect(sock, (struct sockaddr *) &s_in, sizeof(s_in)) == 0,
@@ -75,18 +77,18 @@ void send_command(int sock, const char *format, ...)
 
 
 /* Reads bytes until buffer is full or EOF, returns bytes read. */
-static int read_buffer(int sock, void *buffer, int length)
+static size_t read_buffer(int sock, void *buffer, size_t length)
 {
-    int total = 0;
+    size_t total = 0;
     while (length > 0)
     {
         ssize_t bytes_read = read(sock, buffer, length);
         TEST_SOCK(sock, bytes_read >= 0);
         if (bytes_read == 0)
             break;
-        total += bytes_read;
-        buffer += bytes_read;
-        length -= bytes_read;
+        total  += (size_t) bytes_read;
+        buffer += (size_t) bytes_read;
+        length -= (size_t) bytes_read;
     }
     return total;
 }
@@ -99,7 +101,7 @@ void check_result(int sock)
     if (message[0] != '\0')
     {
         /* Whoops, we have an error message to read. */
-        int bytes_read = read_buffer(sock, message + 1, sizeof(message) - 1);
+        size_t bytes_read = read_buffer(sock, message + 1, sizeof(message) - 1);
         close(sock);
 
         TEST_OK(message[bytes_read] == '\n');
@@ -109,11 +111,13 @@ void check_result(int sock)
 }
 
 
-int fill_buffer(
-    int sock, void *buffer, int sample_size, int buffer_samples, int samples)
+size_t fill_buffer(
+    int sock, void *buffer, size_t sample_size,
+    size_t buffer_samples, size_t samples)
 {
-    int samples_to_read = samples > buffer_samples ? buffer_samples : samples;
-    int bytes_to_read = sample_size * samples_to_read;
+    size_t samples_to_read =
+        samples > buffer_samples ? buffer_samples : samples;
+    size_t bytes_to_read = sample_size * samples_to_read;
     TEST_SOCK_(sock, read_buffer(sock, buffer, bytes_to_read) == bytes_to_read,
         "read", "Unexpected end of input");
     return samples_to_read;
